raa489000: Choose charge port if VBUS present@init
we need to starting sinking VBUS before writing to the
Role control register, otherwise the ASGATE turns off.  This commit
checks to see if VBUS is present on a port at init time and no other
charge port has been selected in order to keep the ASGATE turned on.

BUG=b:198405407
BRANCH=None
TEST=make buildall -j
flash on gooey, verify that VSYS remains up.

Signed-off-by: arthur.lin <arthur.lin@lcfc.corp-partner.google.com>
Change-Id: I5f1ba6b836a66f43564defcaee3c1b0af8dd9e92