zephyr: zmake: Bound size of relative paths for resolve_build_dir tests
In Linux, both filenames and paths are not unbounded in size.  The
typical constants NAME_MAX and PATH_MAX specify the maximum size (in
bytes) of a file name and a file path, respectively.  In Linux, these
constants are 256 and 4096.

The regex which was previously used here assumed file names could be
infinitely long.  Bound it.  Additionally, we need to limit to ASCII
word characters, as Unicode characters may sometimes be more than one
byte.

BUG=b:199769823
BRANCH=none
TEST=run unit tests

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I404302ea10ec27acc9c8e6988bb6dd6050deb09a