zmake: Trim short version on full hashes
The ec_version.h contains both a long build string and also a 31 char
string. Change the short string to end on the the of a hash, i.e
hayato_v2.6.73670-ec:5bd1aa not hayato_v2.6.73670-ec:5bd1aa,os:

Change test to be a real version string with hashes of the right length.

BUG=None
BRANCH=None
TEST=./zephyr/zmake/run_tests.sh

Change-Id: I18d73345036933594b8c74187796c3727f868629
Signed-off-by: Jeremy Bettis <jbettis@google.com>