cr50_fuzz: Add fuzzer for u2f commands
Currently there's only one fuzzer for PinWeaver and one for host
commands in cr50. Add a fuzzer for the u2f commands (generate, sign,
attest) used in the WebAuthn flow to ensure its security. Most regions
of the concerning functions are covered except for pure error code
returns and unreachable regions (currently auth secret is not used in
sign and attest command yet).

BUG=b:172367435
TEST=make host-cr50_u2f_fuzz && \
     ./build/host/cr50_u2f_fuzz/cr50_u2f_fuzz.exe -timeout=10 \
     -ignore_ooms=false -ignore_timeouts=false -fork=71; \
     llvm-profdata merge -sparse default.profraw -o default.profdata; \
     llvm-cov show ./build/host/cr50_u2f_fuzz/cr50_u2f_fuzz.exe \
     -object ./build/host/cr50_u2f_fuzz/RO/board/cr50/dcrypto/u2f.o \
     --instr-profile default.profdata > report

Change-Id: I02b820cf03f7b46ccad7c3bc7b82e73ff45217c6
Signed-off-by: Howard Yang <hcyang@google.com>