ec: Set GCOV tool for cross compile
Use the right gcov tool for cross compile and host builds. The llvm-cov
tool is sort of magical and can handle both llvm and gcc output, but
that is only in the chroot, and so on gitlab we need to pick the right
gcov for each platform.

BRANCH=None
BUG=None
TEST=in docker: make -j CRYPTOC_DIR="${MODULES_DIR}/cryptoc" \
FTDIVERSION=1 HOSTGCOV='gcov' \
CROSS_COMPILE_arm=/opt/zephyr-sdk/arm-zephyr-eabi/bin/arm-zephyr-eabi- \
coverage

Change-Id: Ie95032a4c0706b2bd49574df03e93dcbec7c3eed
Signed-off-by: Jeremy Bettis <jbettis@google.com>