motion_sense: Fix ap_event_interval calculation.
[CONFIG_AP].ec_rate, populated with sysfs entry |hwfifo_timeout|, is used
to calculate |ap_event_interval| the delay before sending a new event to
the AP.
When load_fifo() still existed (CL:938146), it was critical that a
sample was available before gathering the data. When ODR and
|hwfifo_timeout| are closed, |ap_event_interval| is increased by 5%.
Now, since samples are collected after an interrupt triggers due to at
least one element in the sensor FIFO or are scheduled based on the
sensor data rate (forced mode), we are guaranteed to have an event.
We should actually decreate |ap_event_interval| by 5% to be sure an
event is triggered.
When both an interrupt based sensor and a force sensor are at running at
the same ODR, they will probably not be in phase, but there will be
opportunity to send event to the host. One sensor will have a fix delay,
guaranteed to be less than its ODR.

BUG=b:195264765,b:129159505,b:73557414,b:242263033
BRANCH=many.
TEST=Without the change, on gaybrush, we can see that events are sent 2
by 2, event when the AP asks for them immediately.

Using 'iioservice_simpleclient --device_id=3 --channels="timestamp"
--frequency=49 --samples=40' (50Hz ODR):

                  : Before               : After
 Latency tolerance: 0.070408 s           : 0.070408 s
 Max latency      : 0.024089 s           : 0.004047 s
 Min latency      : 0.003339 s           : 0.003167 s
 Median latency   : 0.02319 s            : 0.003565 s
 Mean latency     : 0.013644 s           : 0.003567 s

Change-Id: I9035dd535ff2895be5011335c5f44bd069c9421e
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3172269
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>