TCPMv2: Add DPM setup enum
Currently, the functions to set up VDMs return a VDO count to indicate
success versus an error.  They also have a return of 0, which isn't
strictly defined in the documentation, but is used as an indication a
mode isn't supported in the TBT module at least.

Add a descriptive enum to indicate whether the module set up a message
successfully, has an unsupported mode, or encountered an error.  These
should track roughly to the old positive, zero, and -1 values.

BRANCH=None
BUG=None
TEST=tast typec.Mode*.manual on voxel

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I05fe0ccc5f39e433dd4fbc39c97278b3c1542497