core/cortex-m0: Fix atomic.h compilation with clang
clang warns about the following invalid instructions in the inline assembly:

core/cortex-m0/atomic.h:44:9: error: invalid instruction, any one of the
following would fix this:

<inline asm>:4:1: note: instantiated into assembly here
orr r3, r3, r2
^
<inline asm>:4:1: note: instantiated into assembly here
bic r2, r2, r1
^

Based on
https://developer.arm.com/documentation/dui0489/i/arm-and-thumb-instructions/orr
and
https://developer.arm.com/documentation/dui0489/i/arm-and-thumb-instructions/bic,
it appears that when using Thumb "orr" and "bic" can only be used inside an IT
block. We're not in an IT block, so we need to use the "orrs" and "bics"
version which can only be used outside an IT block. The "s" indicates that the
condition flags will be updated, but we already indicate that we are clobbering
condition codes with "cc" in the inline asm.

Also see https://developer.arm.com/documentation/ddi0432/c/CHDCICDF,
which lists the Cortex-M0 instruction set.

BRANCH=none
BUG=b:172020503
TEST=none

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ifafaf67080e25ab2c5d2aad3efc90a9d61978bf9