TCPMv2: Handle timing corner case in PRL discard
If a message is received by the PD_INT task between the PRL_RX run and
PRL_TX run, we may end up sending a message to the TCPC even though the
PRL_RX will be forcing a discard next PRL cycle.  This will get the PE
out of sync, as it will assume the message was discarded when it was in
fact sent.

Peek into our receive queue right before sending a message to the TCPC,
in order to reduce the liklihood of a mix-up like this happening.

BRANCH=None
BUG=None
TEST=on voxel, run tast typec.Mode*.manual and observe we can
successfully exit DP mode even when a VDM:Attention command comes in
mid-PRL cycle

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I6979bf53972e40794dfae83f631b865a00d73302