core/cortex-m[0]: Work around clang inline-asm warning
clang warns that we're clobbering a reserved register:

inline asm clobber list contains reserved registers: R7
[-Werror,-Winline-asm].

According to
https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Clobbers-and-Scratch-Registers:

When the compiler selects which registers to use to represent input and
output operands, it does not use any of the clobbered registers.

Before this change, the only remaining registers for the compiler to use
to hold the values represented by the symbolic names "pregs" and
"pstack" are R0 and R12. After this change, this still holds on clang,
since it won't clobber the reserved register R7 even if it's not on the
clobber list.

BRANCH=none
BUG=b:172020503
TEST=make V=1 CC=arm-none-eabi-clang BOARD=elm
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I5182ef24f4465d68c97f4160a45cd479fe52e017