core/minute-ia: Fix inline asm
clang produces the following warning:

core/minute-ia/panic.c:119:19: error: variable 'eax' is uninitialized when used
here [-Werror,-Wuninitialized]
        pdata->x86.eax = eax;
                         ^~~
core/minute-ia/panic.c:109:23: note: initialize the variable 'eax' to silence
this warning
        register uint32_t eax asm("eax");

When specifying registers for local variables they need to be specified as
constraints on an asm block:
https://gcc.gnu.org/onlinedocs/gcc/Local-Register-Variables.html#Local-Register-Variables
In this case we simply use an empty asm block since we're just trying to
extract the values of the registers.

As indicated by the TEST line, the output is identical before and after
this change.

BRANCH=none
BUG=b:172020503
TEST=make CC=clang V=1 BOARD=arcada_ish
TEST=./util/compare_build.sh -b all -j 70
      => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I6fbd7061d9af734f02ec3dbb1329cdffac5e2a7d