cherry: powercycle VBUS after AP XHCI done
Implement VBUS cycle according to USB Type-C Spec R2.0:

10.3.1 Hub Downstream Facing Port State Descriptions

10.3.1.1 DSPORT.Powered-off

A hub may provide power to its downstream ports all of the time to support
power applications from a USB port. Such hubs must ensure that Enhanced
SuperSpeed devices on its downstream-facing ports attempt Enhanced
SuperSpeed connection once upstream VBUS is seen by the hub. The
recommended method to achieve this is to cycle VBUS off for a duration or
by actively discharging so that it is seen to be off by the downstream
device.

BUG=b:200772411
TEST=manually
BRANCH=main

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: I674391622698b6e241b0a97063b3abbc8550e365