driver: isl923x: Fix prochot for AC/DC
The datasheet for the isl923x specifies that the prochot register values
do not take into account the resistor (which *CURRENT_TO_REG do). This
is likely a bug in the code which was never seed due to board
configurations matching the recommended resistor values.

Replace *CURRENT_TO_REG calls with the appropriate masking and document
the reason.

BRANCH=none
BUG=b:201602829,b:201819565
TEST=make buildall -j

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: If13114d5ce985c8a144469fdea59715023e380ee