zephyr: test: ln9310 fix emulator and newer chip powers_up tests
Due to a misunderstanding of the ln9310 software_enable function doc,
we were mistakenly testing that calling the ln9310_software_enable was
unnecessary to power up new versions of the chip.

This commit fixes the ln9310 emulator and driver tests to require
the ln9310_software_enable function in to power up the ln9310.

BRANCH=none
BUG=b:184856083
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Aaron Massey <aaronmassey@chromium.org>

Cq-Depend: chromium:3200519
Change-Id: I5c5ddcc46342203298655342cc8b1365263d4f3b