tree: Add __keep to interrupt handlers
DECLARE_IRQ creates a declaration of the function with the "used"
attribute. The declaration with this attribute must come before the
definition of the function or the compiler can ignore it. If it ignores
the "used" attribute, the linker can optimize it out when doing LTO:

ld.lld: error: undefined symbol: adc_interrupt
>>> referenced by adc.c:357 (platform/ec/chip/npcx/adc.c:357)
>>>               lto.tmp:(irq_10_handler)

ld.lld: error: undefined symbol: __gpio_rtc_interrupt
>>> referenced by gpio-npcx7.c:171 (platform/ec/chip/npcx/gpio-npcx7.c:171)
>>>               lto.tmp:(irq_7_handler)

A few changes in chip/stm32 are excluded in this commit since they
create a difference in the compilation output. A separate commit will
address those.

BRANCH=none
BUG=b:172020503
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ie184158c69333309232299eb1ed54a18d25c4352