gitlab: Use cache to only get source code once
There was a recent breakage where one job failed because of a git clone
timeout. But more importantly, we should make sure all the jobs are
using the same commit.

Change CI jobs to use the gitlab cache to keep the non-ec git repos so
only the first job (seed_cache) will do the git clone, and the others
will just pull the files from the cache.

The cache key is the ec commit sha that is being built.

The total build time after this is 40m, and the builds before this were
37-39m, so it doesn't seem to be a change.

BUG=None
BRANCH=None
TEST=Pushed to coverage branch on gitlab:
https://gitlab.com/zephyr-ec/ec/-/pipelines/382237078

Change-Id: Ic596464fbf90c2964075b4f8f3e6bdc77bde531b
Signed-off-by: Jeremy Bettis <jbettis@google.com>