fips: silence fips section filler
Script inserting FIPS checksum into the image uses the dd utility
which generates stderr output even when there is no errors.

This patch adds code which captures the dd stderr output and prints it
out only if there is an actual error. stdout output of the script is
suppressed unless make was invoked with V=1.

Also made a few modifications as requested by shellcheck.

BUG=none
TEST=make output does not have extra lines.

     built and ran a Cr50 image, it reports successful FIPS integrity
     self check.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I9121bc5a9a40633b9a3d18ea5766bc1ed274a9c2