gsctool: support the new AP RO status VC format
Shimless RMA needs to tell if the button combo was triggered even if the
board doesn't support verifying the AP RO hash.

The cr50 vendor command will now return the key combo status in addition
to the AP RO verification status. This change modifies gsctool to print
the AP RO verification status and the key combo status.

The old vendor command result
- result[0] is the AP RO verification status [PASS, FAIL, NOT_RUN
  UNSUPPORTED]

The new vendor command result
- result[0] is the AP RO verification status [PASS, FAIL, NOT_RUN
  UNSUPPORTED]
- result[1] is the combo status 0 for not pressed. 1 for pressed this
  boot.

gsctool can determine the button status for three of the results from
the old vendor command. This change will convert the result of the old
vendor command into yes, no, or "unknown"
- "yes" - PASS and FAIL mean the combo was pressed
- "no" - NOT_RUN means the combo wasn't pressed
- "unknown" - gsctool can't determine the button state from UNSUPPORTED

BUG=b:181000999
TEST=use gsctool to get the AP RO verification status on cr50 images
with the new and old version of the get AP RO status vendor command.

Change-Id: Ic8bc33e0019b4d4924f784ca14e597ed2bd7dd91
Signed-off-by: Mary Ruthven <mruthven@chromium.org>