common/console: No PROMPT if CONFIG_EXPERIMENTAL_CONSOLE
CONFIG_EXPERIMENTAL_CONSOLE shift responsibility of command line
editing and history away from the EC.  As a part of that change, the
EC should not emit the prompt character ">".  It seems that every
instance but one had been enclosed in ifdef, this CL will take care of
that one, as well as make sure that PROMPT is not defined at all when
CONFIG_EXPERIMENTAL_CONSOLE is enabled, allowing the compiler to
prevent any regressions.

BUG=none
BRANCH=none
TEST=build servo_micro with CONFIG_EXPERIMENTAL_CONSOLE enabled

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ifdab902cf877c7adee3d8e564cd234ffb19e317b