common/console: No PROMPT if CONFIG_EXPERIMENTAL_CONSOLE
The existing CONFIG_EXPERIMENTAL_CONSOLE shifts responsibility of
command line editing and history away from the EC.  As a part of that
change, the EC should not emit the prompt character ">".  It seems
that every instance but one had already been enclosed in ifdef.  This
CL will take care of that last one, as well as make sure that PROMPT
and CTRL are not defined when CONFIG_EXPERIMENTAL_CONSOLE is enabled,
allowing the compiler to prevent any regressions.

BUG=none
BRANCH=none
TEST=build servo_micro with CONFIG_EXPERIMENTAL_CONSOLE enabled

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ifdab902cf877c7adee3d8e564cd234ffb19e317b