core/cortex-m0: Free up r3 in exception_panic
According to
https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Clobbers-and-Scratch-Registers:

When the compiler selects which registers to use to represent input and
output operands, it does not use any of the clobbered registers.

When compiling the inline assembly, clang reports the error:

inline assembly requires more register than available

This appears to be true for gcc as well, but instead of warning, it just
silently clobbers r3. For clang, remove r3 from the clobber list so that
it can use it.

BRANCH=none
BUG=b:172020503
TEST=make CC=arm-none-eabi-clang BOARD=c2d2

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: If9317dc9e4cda4a806ce4ed1b6d05723bef9869b