config: Add CONFIG_BATTERY_NO_AUTO_DETECT option
There are some battery types which cannot be identified
by normal manufacturing name or device name. For devices,
which uses this type of batteries can be configured
with CONFIG_BATTERY_NO_AUTO_DETECT to indicate the same.
Based on this config, the idea is to set the correct
battery type at runtime and then use the manufacturing or
device name to identify the battery later. The motivation
of this change is to have a single binary for multiple
platforms that have different battery configuration.

BUG=b:194163586
BRANCH=none
TEST=make buildall -j

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I488ca3d72250311630724381a8cfeb52d8dad657